adjust bj card define

+++ Server/GBiJiServer/src/BiJi/BJPlayerCard.h
@@ -15,14 +15,14 @@ public:
 	public:
 		stGroupCard() { m_vCard.resize(PEER_CARD_COUNT); }
 		void reset();
-		void setCard( std::vector<uint16_t>& vCompsitCard );
+		void setCard( std::vector<uint8_t>& vCompsitCard );
 		eBJCardType getType() { return m_eCardType; }
 		uint32_t getWeight() { return m_nWeight; }
-		uint16_t getCardByIdx( uint8_t nIdx );
+		uint8_t getCardByIdx( uint8_t nIdx );
 	protected:
 		eBJCardType m_eCardType;
 		uint32_t m_nWeight;
-		std::vector<uint16_t> m_vCard;  // uint16_t , 两个8位，前一个8位表示赖子牌，后面的8位表示癞子表示牌。如果没有癞子信息，前一个8位都是0 ；
+		std::vector<uint8_t> m_vCard;
 	};
 public:
 	BJPlayerCard();
@@ -32,10 +32,10 @@ public:
 	uint32_t getWeight()override;
 	IPeerCard* swap(IPeerCard* pTarget)override;
 	uint8_t getCardByIdx(uint8_t nidx)override;
-	bool getGroupInfo( uint8_t nGroupIdx , uint8_t& nGroupType , std::vector<uint16_t>& vGroupCards ); // uint16_t , 两个8位，前一个8位表示赖子牌，后面的8位表示癞子表示牌。如果没有癞子信息，前一个8位都是0 ；
+	bool getGroupInfo( uint8_t nGroupIdx , uint8_t& nGroupType , std::vector<uint8_t>& vGroupCards );
 	void setCurGroupIdx( uint8_t nGrpIdx );
 	uint8_t getCurGroupIdx();
-	bool setCardsGroup(std::vector<uint16_t>& vGroupedCards); // uint16_t , 两个8位，前一个8位表示赖子牌，后面的8位表示癞子表示牌。如果没有癞子信息，前一个8位都是0 ；
+	bool setCardsGroup(std::vector<uint8_t>& vGroupedCards);
 	bool holdCardToJson(Json::Value& vHoldCards);
 	bool groupCardToJson(Json::Value& vHoldCards);
 	// xi pai used 
@@ -43,7 +43,7 @@ public:
 	stGroupCard& getGroupByIdx(uint8_t nGroupIdx );
 	eXiPaiType getXiPaiType( bool isEnableSanQing, bool isEnableShunQingDaTou );
 protected:
-	void autoChoseGroup(std::vector<uint16_t>& vGroupedCards);
+	void autoChoseGroup(std::vector<uint8_t>& vGroupedCards);
 protected:
 	std::vector<uint8_t> m_vHoldCards;
 	std::vector<stGroupCard> m_vGroups;
